Player Respawn 1.0
Added a debug mode to kill the player if they click "X"

Player respawns on ship if either they or the ship dies.

Added map borders (jank, dont know how to make less jank). Player health drains if they are outside the borders

Bugs:
If the player dies while the whale is close enough to the borders, nothing is stopping the player from spawning out of bounds

If the player dies while on the ship helm, they are permanently stuck on the ship helm (the same might be true for turret and/or processor, did not check)