Harpoon part 1
Harpoon fires, attaches to whatever it hits (excluding the ship), and detaches after a certain amount of time

Harpoon and Ship move with the object the harpoon had hit. Player does not yet.

Commented out a line in the TurretInteractable script that was disabling the FirstPersonController script. This line was making the player not move with the ship properly. However, commenting it out made the player not return to their original position after leaving the turret.