Harpoon
Harpoon fires and travels until either it reaches max distance, or hits the whale's hitbox

If harpoon hits the whale, it latches onto it. The ship moves closer to the harpoon if it is further than the harpoon's max distance. The ship moves closer to the harpoon at the same speed the whale is currently moving at, so if the ship is max distance from the harpoon, it should stay max distance unless the whale starts approaching them. The player is not allowed to move the ship on their own until the harpoon has detached from the whale.

After the harpoon has been out for its lifetime, it returns to the turret.